+++ src/main/java/ru/hogwarts/school/repository/StudentRepository.java
@@ -6,5 +6,5 @@ import ru.hogwarts.school.model.Student;
 import java.util.Collection;
 
 public interface StudentRepository extends JpaRepository<Student,Long> {
-    Collection<Student> findByAge(Long age);
+    Collection<Student> findByAge(int age);
 }
